requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
google-cloud-firestore==2.13.1
python-dotenv==1.0.0
